012-functions-env/code/requirements.txt
functions-framework
google-cloud-core
